+++ js/main.js
@@ -1,16 +1,5 @@
-import { map, initData } from './map.js';
-import { setActiveState, setActiveStateFormMapFilters } from './form.js';
-import { getData } from './api.js';
-import { displayError } from './error.js';
+import { initMap } from './map.js';
+import { setInactiveState } from './form.js';
 
-const onLoadData = (noticeList) => {
-  setActiveStateFormMapFilters();
-  initData(noticeList);
-};
-
-const onLoadMap = () => {
-  setActiveState(map);
-  getData(onLoadData, displayError);
-};
-
-map.whenReady(onLoadMap);
+setInactiveState();
+initMap();
