+++ js/map.js
@@ -1,5 +1,8 @@
-import { setInactiveState, setAddressInput } from './form.js';
+import { setAddressInput, setActiveState } from './form.js';
+import { setActiveStateFormMapFilters } from './form-filters.js';
 import { createCardElement } from './template.js';
+import { displayError } from './error.js';
+import { getData } from './api.js';
 
 const MAXIMUM_DISPLAY_NOTICE = 10;
 
@@ -15,18 +18,18 @@ const INIT_POINT = {
   Zoom: 10,
 };
 
-setInactiveState();
-
 const map = L.map('map-canvas')
   .setView(INIT_POINT.LatLng, INIT_POINT.Zoom);
 
-L.tileLayer(
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-  {
-    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-  },
-)
-  .addTo(map);
+const loadOpenstreetMap = () => {
+  L.tileLayer(
+    'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+    {
+      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+    },
+  )
+    .addTo(map);
+};
 
 const mainPinIcon = L.icon({
   iconUrl: '../img/main-pin.svg',
@@ -66,14 +69,16 @@ const onAddressInput = (evt) => {
   setAddressInput(evt.target.getLatLng());
 };
 
-addressMarker.addTo(map);
-addressMarker.addEventListener('moveend', onAddressInput);
+const initAddressMarker = () => {
+  addressMarker.addTo(map);
+  addressMarker.addEventListener('moveend', onAddressInput);
+};
 
 const setMapView = () => {
   map.setView(INIT_POINT.LatLng, INIT_POINT.Zoom);
 };
 
-const initAddressMarker = () => {
+const resetAddressMarker = () => {
   setMapView();
   addressMarker.setLatLng(INIT_POINT.LatLng);
   setAddressInput(addressMarker.getLatLng());
@@ -121,11 +126,29 @@ const initData = (noticeList) => {
   displayInitData();
 };
 
+const onLoadData = (noticeList) => {
+  setActiveStateFormMapFilters();
+  initData(noticeList);
+};
+
+const onLoadMap = () => {
+  setActiveState();
+  getData(onLoadData, displayError);
+};
+
+const initMap = () => {
+  loadOpenstreetMap();
+  initAddressMarker();
+  map.whenReady(onLoadMap);
+};
+
+const getMap = () => map;
+
 export {
-  map,
-  initData,
+  getMap,
+  initMap,
   displayInitData,
-  initAddressMarker,
+  resetAddressMarker,
   closeAllPopup,
   displaySelectedMarkerList,
   getMarkerDataList
